Add Dockerfile.
Add a Dockerfile that just wraps the code into a Debian 12 Bookworm node.js container and sets it as the entrypoint.
Add instructions to the README on building the docker image manually, and running it as a one-shot container.
Fix a mistake in the subcommands in the README.
Add some additional detail to the fullMigrate command about what folders are created.